--- src/types/pcl-js.d.ts
@@ -1,4 +1,0 @@
-// declare module "pcl.js/PointCloudViewer" {
-//   import PointCloudViewer from "pcl.js/dist/visualization/PointCloudViewer";
-//   export default PointCloudViewer;
-// }
